% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIDs.R
\name{checkGPL}
\alias{checkGPL}
\title{Check the input gpl if in our platform list}
\usage{
checkGPL(gplnum)
}
\arguments{
\item{gplnum}{GPL(GEO platform) number, eg: GPL570}
}
\value{
returns a boolean value
}
\description{
Check the input gpl if in our platform list
}
