% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterExpressionMatrix.R
\name{filterEM}
\alias{filterEM}
\title{Filter expression matrix}
\usage{
filterEM(eMatrix, probe2gene, omitNA = T)
}
\arguments{
\item{eMatrix}{is an expression matrix which rownames are probe ids and every column is a sample}

\item{probe2gene}{is a data frame which includes probe ids and corresponding gene Symbols}

\item{omitNA}{omit NA symbol rows in matrix}
}
\value{
filter ids expression ma
}
\description{
\code{filterEM} returns
}
\examples{
eMatrix <- system.file("extdata", "matrixData.Rdata", package="IDmap")
allMapIds <- system.file("extdata", "probe2gene.Rdata", package="IDmap")
load(eMatrix)
load(allMapIds)
## head(matrixData)
## head(probe2gene)
res <- filterEM(matrixData,probe2gene,omitNA=T)
}
