% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGPLsoft.R
\name{get_geo}
\alias{get_geo}
\title{Download probe annotations from GEO}
\usage{
get_geo(GPL, destDir = getwd())
}
\arguments{
\item{GPL, destDir}{}
}
\value{
{getGPLsoft} will download probe annotations from GEO ,then the output will be a dataframe
}
\description{
\code{getGPLsoft} download probe annotation from GEO database for the input GPL Accession Number
}
\examples{
getGPLsoft("GPL13607", destdir=getwd())

}
