% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIDs.R
\name{getAnno}
\alias{getAnno}
\title{Get Annotation}
\usage{
getAnno(gplnum, source = "pipe", biotype = "protein_coding",
  lncRNA = F)
}
\arguments{
\item{gplnum, source, biotype, lncRNA}{}
}
\value{

}
\description{
\code{getAnno} returns annotations for target gpl
}
\examples{
humanAnno <- getAnno(gplnum,source, biotype,lncRNA)
head(humanAnno)
}
