% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIDs.R
\name{getGPLinfo}
\alias{getGPLinfo}
\title{Get GPL info}
\usage{
getGPLinfo(gplnum)
}
\arguments{
\item{gplnum}{GPL(GEO platform) number, eg: GPL570}
}
\value{
detail information of the GEO platform
}
\description{
Get GPL info
}
