% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIDs.R
\name{probeIdmap}
\alias{probeIdmap}
\title{Mapping probeIds to gene symbols}
\usage{
probeIdmap(probeids, datasets, probeIdcol)
}
\arguments{
\item{probeids}{input probe ids}

\item{datasets}{a dataframe which contains probe annotaions}

\item{probeIdcol}{the column name of probe ids}
}
\value{
a dataframe of gene ids mapping to probe ids
}
\description{
\code{probeIdmap} returns a list of gene ids for the input probe ids
}
\examples{
probeids <- c("A_23_P101521","A_33_P3695548","A_33_P3266889","A_33_P3266886")
gplnum <- "GPL10332"
source <- "pipe"
biotype <- "protein_coding"
lncRNA <- FALSE
datasets <- getAnno(gplnum,source, biotype,lncRNA)
mapRes <- probeIdmap(probeids, datasets)
head(mapRes)
}
