% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveRdata.R
\name{refreshRdata}
\alias{refreshRdata}
\title{Save the Rdata into our package}
\usage{
refreshRdata(refresh = F)
}
\arguments{
\item{refresh}{Don't use this function if not neccessary}
}
\description{
So far, we just have 6 files
}
\keyword{Rdata}
\keyword{refresh}
