% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIDs.R
\name{probeIdmap}
\alias{probeIdmap}
\title{Mapping probeIds to gene symbols}
\usage{
probeIdmap(probeids, datasets)
}
\arguments{
\item{probeids, datasets}{}
}
\value{
if the input dataset is in our platform list(see our manual) ,then the output will be a dataframe,
which includes a list of gene ids mapping to probe ids
}
\description{
\code{probeIdmap} returns a list of gene ids for the input probe ids
}
